Fixes bug with add item and pulls uploadImage into a util function for re-use
